+++ python/cudf/cudf/tests/test_dataframe.py
@@ -2666,27 +2666,77 @@ def test_column_assignment():
 
 def test_select_dtype():
     gdf = gd.datasets.randomdata(
-        nrows=20, dtypes={"a": "category", "b": int, "c": float}
+        nrows=20, dtypes={"a": "category", "b": int, "c": float, "d": str}
     )
+    pdf = gdf.to_pandas()
 
-    assert_eq(gdf[["c"]], gdf.select_dtypes("float64"))
-    assert_eq(gdf[["c"]], gdf.select_dtypes(np.float64))
-    assert_eq(gdf[["c"]], gdf.select_dtypes(include=["float64"]))
+    assert_eq(pdf.select_dtypes("float64"), gdf.select_dtypes("float64"))
+    assert_eq(pdf.select_dtypes(np.float64), gdf.select_dtypes(np.float64))
+    assert_eq(pdf.select_dtypes(include=["float64"]),
+              gdf.select_dtypes(include=["float64"]))
+    assert_eq(pdf.select_dtypes(include=['object', 'int', 'category']),
+              gdf.select_dtypes(include=['object', 'int', 'category']))
 
-    assert_eq(gdf[["b", "c"]], gdf.select_dtypes(include=["int64", "float64"]))
-    assert_eq(gdf[["b", "c"]], gdf.select_dtypes(include=np.number))
+    assert_eq(pdf.select_dtypes(include=["int64", "float64"]),
+              gdf.select_dtypes(include=["int64", "float64"]))
+    assert_eq(pdf.select_dtypes(include=np.number),
+              gdf.select_dtypes(include=np.number))
     assert_eq(
-        gdf[["b", "c"]], gdf.select_dtypes(include=[np.int64, np.float64])
+        pdf.select_dtypes(include=[np.int64, np.float64]),
+        gdf.select_dtypes(include=[np.int64, np.float64])
     )
 
-    assert_eq(gdf[["a"]], gdf.select_dtypes(include=["category"]))
-    assert_eq(gdf[["a"]], gdf.select_dtypes(exclude=np.number))
+    assert_eq(pdf.select_dtypes(include=["category"]),
+              gdf.select_dtypes(include=["category"]))
+    assert_eq(pdf.select_dtypes(exclude=np.number),
+              gdf.select_dtypes(exclude=np.number))
 
     with pytest.raises(TypeError):
-        assert_eq(gdf[["a"]], gdf.select_dtypes(include=["Foo"]))
+        assert_eq(pdf.select_dtypes(include=["Foo"]),
+                  gdf.select_dtypes(include=["Foo"]))
 
     with pytest.raises(ValueError):
         gdf.select_dtypes(exclude=np.number, include=np.number)
+    with pytest.raises(ValueError):
+        pdf.select_dtypes(exclude=np.number, include=np.number)
+
+    gdf = DataFrame({'A': [3, 4, 5],
+                     'C': [1, 2, 3],
+                     'D': ['a', 'b', 'c']})
+    pdf = gdf.to_pandas()
+    assert_eq(pdf.select_dtypes(include=['object', 'int', 'category']),
+              gdf.select_dtypes(include=['object', 'int', 'category']))
+    assert_eq(pdf.select_dtypes(include=['object'], exclude=['category']),
+              gdf.select_dtypes(include=['object'], exclude=['category']))
+
+    gdf = gd.DataFrame({'a': range(10), 'b': range(10, 20)})
+    pdf = gdf.to_pandas()
+    assert_eq(pdf.select_dtypes(include=['category']),
+              gdf.select_dtypes(include=['category']))
+    assert_eq(pdf.select_dtypes(include=['float']),
+              gdf.select_dtypes(include=['float']))
+    assert_eq(pdf.select_dtypes(include=['object']),
+              gdf.select_dtypes(include=['object']))
+    assert_eq(pdf.select_dtypes(include=['int']),
+              gdf.select_dtypes(include=['int']))
+    assert_eq(pdf.select_dtypes(exclude=['float']),
+              gdf.select_dtypes(exclude=['float']))
+    assert_eq(pdf.select_dtypes(exclude=['object']),
+              gdf.select_dtypes(exclude=['object']))
+    assert_eq(pdf.select_dtypes(include=['int'], exclude=['object']),
+              gdf.select_dtypes(include=['int'], exclude=['object']))
+    with pytest.raises(ValueError):
+        pdf.select_dtypes()
+    with pytest.raises(ValueError):
+        gdf.select_dtypes()
+
+    gdf = gd.DataFrame({'a': gd.Series([], dtype='int'),
+                        'b': gd.Series([], dtype='str')})
+    pdf = gdf.to_pandas()
+    assert_eq(pdf.select_dtypes(exclude=['object']),
+              gdf.select_dtypes(exclude=['object']))
+    assert_eq(pdf.select_dtypes(include=['int'], exclude=['object']),
+              gdf.select_dtypes(include=['int'], exclude=['object']))
 
 
 def test_select_dtype_datetime():