+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 - PR #2168 Use cudf.Column for CategoricalColumn's categories instead of a tuple
 - PR #2193 Added more docuemtnation to `type_dispatcher` for specializing dispatched functors
 - PR #2197 CSV Writer: Expose `chunksize` as a parameter for `to_csv`
+- PR #2209 Matching `get_dummies` & `select_dtypes` behavior to pandas
 
 ## Bug Fixes